+++ concordium-cis2/src/lib.rs
@@ -1022,6 +1022,33 @@ where
     fn from(err: TransferError) -> Self { Cis2Error::Custom(X::from(err)) }
 }
 
+impl<X> From<UpgradeError> for Cis2Error<X>
+where
+    X: From<UpgradeError>,
+{
+    #[inline]
+    /// Converts the error by wrapping it in [Self::Custom].
+    fn from(err: UpgradeError) -> Self { Cis2Error::Custom(X::from(err)) }
+}
+
+impl<X> From<QueryAccountBalanceError> for Cis2Error<X>
+where
+    X: From<QueryAccountBalanceError>,
+{
+    #[inline]
+    /// Converts the error by wrapping it in [Self::Custom].
+    fn from(err: QueryAccountBalanceError) -> Self { Cis2Error::Custom(X::from(err)) }
+}
+
+impl<X> From<QueryContractBalanceError> for Cis2Error<X>
+where
+    X: From<QueryContractBalanceError>,
+{
+    #[inline]
+    /// Converts the error by wrapping it in [Self::Custom].
+    fn from(err: QueryContractBalanceError) -> Self { Cis2Error::Custom(X::from(err)) }
+}
+
 impl<X> From<NewReceiveNameError> for Cis2Error<X>
 where
     X: From<NewReceiveNameError>,